Re-use load_config() functions
