+++ pyseasonal/aggregate_hindcast.py
@@ -13,10 +13,9 @@ import xarray as xr
 import os
 import pandas as pd
 import pdb #then type <pdb.set_trace()> at a given line in the code below
-import yaml
-from pathlib import Path
 
-from functions_seasonal import transform_gcm_variable
+from pyseasonal.config import load_config
+from pyseasonal.functions_seasonal import transform_gcm_variable
 
 # INDICATE CONFIGURATION FILE ######################################
 
@@ -26,26 +25,9 @@ configuration_file = 'config_for_aggregate_hindcast_Canarias.yaml'
 
 ####################################################################
 
-#this is a function to load the configuration file
-def load_config(config_file='config/'+configuration_file):
-    """Load configuration from YAML file"""
-    config_path = Path(__file__).parent.parent / config_file
-    print('The path of the configuration file is '+str(config_path))
-    with open(config_path, 'r') as file:
-        config = yaml.safe_load(file)
-
-    # Setup paths based on GCM_STORE environment variable
-    gcm_store = os.getenv('GCM_STORE', 'lustre')
-    if gcm_store in config['paths']:
-        paths = config['paths'][gcm_store]
-        config['paths'] = paths
-    else:
-        raise ValueError('Unknown entry for <gcm_store> !')
-
-    return config
 
 # Load configuration
-config = load_config()
+config = load_config(configuration_file)
 
 # Extract configuration variables
 model = config['model']
