+++ pyseasonal/products/seas2ipe.py
@@ -10,34 +10,8 @@ import pandas as pd
 import dask
 import sys
 import pdb
-import yaml
-from pathlib import Path
 
-
-def load_config(config_file):
-    """Load configuration from YAML file"""
-    config_path = Path(__file__).parent.parent / config_file
-    print('The path of the configuration file is '+str(config_path))
-    with open(config_path, 'r') as file:
-        config = yaml.safe_load(file)
-
-    # Setup paths based on GCM_STORE environment variable
-    gcm_store = os.getenv('GCM_STORE', 'lustre')
-    if gcm_store in config['paths']:
-        paths = config['paths'][gcm_store]
-        # Handle special cases for argo environment
-        if gcm_store == 'argo':
-            data_dir = os.getenv("DATA_DIR", "")
-            paths['home'] = data_dir
-            paths['path_gcm_base'] = data_dir + paths['path_gcm_base']
-            paths['path_gcm_base_derived'] = data_dir + paths['path_gcm_base_derived']
-            paths['path_gcm_base_masked'] = data_dir + paths['path_gcm_base_masked']
-            paths['dir_forecast'] = data_dir + paths['dir_forecast']
-        config['paths'] = paths
-    else:
-        raise Exception(f'ERROR: unknown entry for <gcm_store> !')
-
-    return config
+from pyseasonal.config import load_config_argo
 
 
 # CREATE LIST CONTAINING ALL DOMAINS TO BE PROCESSED ######################################
@@ -70,7 +44,7 @@ for do in np.arange(len(domain_list)):
 
 
     # Load configuration
-    config = load_config(configuration_file)
+    config = load_config_argo(configuration_file)
 
     # # Example year and run to run without passing any input arguments; comment or delete the next two lines in operative use
     # year_init = 2024 #a list containing the years the forecast are initialized on, will be looped through with yy
